Все выношу в отдельный бин
