+++ 2020-11/spring-02/spring-02-exercise/src/main/java/ru/otus/spring/service/Questionnaire.java
@@ -1,0 +1,7 @@
+package ru.otus.spring.service;
+
+import java.io.IOException;
+
+public interface Questionnaire {
+    void QuestionnaireExec() throws IOException;
+}
